# modules/apps/headless/headless-portal-instances/headless-portal-instances-impl/rest-openapi.yaml
components:
    schemas:
        PortalInstance:
            # @review
            properties:
                active:
                    readOnly: true
                    type: boolean
                companyId:
                    description:
                        internal unique key.
                    format: int64
                    type: integer
                domain:
                    description:
                        domain used for email authentication.
                    type: string
                portalInstanceId:
                    description:
                        public unique key (corresponds to company's webId field)
                    type: string
                virtualHost:
                    type: string
            type: object
info:
    description:
        ""
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    title: Liferay Portal Instances Headless API
    version: v1.0
openapi: 3.0.1
paths:
    /portal-instances:
        get:
            description:
                Retrieves the portal instances
            operationId: getPortalInstancesPage
            parameters:
                - in: query
                  explode: true
                  name: skipDefault
                  required: false
                  schema:
                      type: boolean
                  style: form
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/PortalInstance"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/PortalInstance"
                                type: array
                    description:
                        default response
            # @review
            tags: ["PortalInstance"]
        post:
            description:
                Adds a new portal instance
            operationId: postPortalInstance
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PortalInstance"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/PortalInstance"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                    description:
                        default response
            # @review
            tags: ["PortalInstance"]
    /portal-instances/{portalInstanceId}:
        delete:
            description:
                Removes the portal instance
            operationId: deletePortalInstance
            parameters:
                - in: path
                  name: portalInstanceId
                  required: true
                  schema:
                      type: string
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            # @review
            tags: ["PortalInstance"]
        get:
            description:
                Retrieves the portal instance
            operationId: getPortalInstance
            parameters:
                - in: path
                  name: portalInstanceId
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                    description:
                        default response
            # @review
            tags: ["PortalInstance"]
        patch:
            # @review
            description:
                Updates the portal instance with information sent in the request body. Only the provided fields are
                updated.
            operationId: patchPortalInstance
            parameters:
                - in: path
                  name: portalInstanceId
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PortalInstance"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/PortalInstance"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/PortalInstance"
                    # @review
                    description:
                        "Portal instance successfully updated"
            # @review
            tags: ["PortalInstance"]
    /portal-instances/{portalInstanceId}/activate:
        # @review
        put:
            description:
                Activates the portal instance
            operationId: putPortalInstanceActivate
            parameters:
                - in: path
                  name: portalInstanceId
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        default response
            tags: ["PortalInstance"]
    /portal-instances/{portalInstanceId}/deactivate:
        # @review
        put:
            description:
                Deactivates the portal instance. When a portal instance is deactivated, its virtual host will not longer
                respond requests.
            operationId: putPortalInstanceDeactivate
            parameters:
                - in: path
                  name: portalInstanceId
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        default response
            tags: ["PortalInstance"]